add userController
